% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{chrome_picker}
\alias{chrome_picker}
\title{Chrome Picker Input}
\usage{
chrome_picker(input_id, label, color = "#112446",
  disable_alpha = FALSE, inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{disable_alpha}{Remove alpha slider and options from picker.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Chrome Picker Input
}
\examples{
# See picker in Viewer (or browser)
chrome_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {

  library(shiny)

  ui <- fluidPage(
    tags$h2("Chrome picker color"),
    chrome_picker("inline", "Pick!", inline = TRUE),
    verbatimTextOutput("res_inline"),
    chrome_picker("btn", "Pick!", inline = FALSE),
    verbatimTextOutput("res_btn")
  )

  server <- function(input, output, session) {
    output$res_inline <- renderPrint(input$inline)
    output$res_btn <- renderPrint(input$btn)
  }

  shinyApp(ui, server)

}
}
