% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{circle_picker}
\alias{circle_picker}
\title{Circle Picker Input}
\usage{
circle_picker(input_id, label, color = "#112446", colors = NULL,
  circle_size = 28, circle_spacing = 14, inline = FALSE,
  width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{colors}{Vector of colors to include in the widget.}

\item{circle_size}{Number, Value for circle size.}

\item{circle_spacing}{Number, Value for spacing between circles.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Circle Picker Input
}
\examples{
# See picker in Viewer (or browser)
circle_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {



}
}
