% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker.R
\name{color_picker}
\alias{color_picker}
\title{Color Picker Inline}
\usage{
color_picker(inputId, color = "#FFFFFF", picker = c("SketchPicker",
  "BlockPicker", "ChromePicker", "CirclePicker", "CompactPicker",
  "GithubPicker", "HuePicker", "MaterialPicker", "PhotoshopPicker",
  "SliderPicker", "SwatchesPicker", "TwitterPicker", "AlphaPicker"),
  width = NULL, height = NULL)
}
\arguments{
\item{inputId}{Widget id to retrieve value server-side.}

\item{color}{Default color.}

\item{picker}{Style of the picker.}

\item{width, height}{Width and height for the widget.}
}
\description{
Create a color picker among 13 available styles.
}
