% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker_btn.R
\name{color_picker_btn}
\alias{color_picker_btn}
\title{Color Picker with Button}
\usage{
color_picker_btn(..., picker = c("SketchPicker", "BlockPicker",
  "ChromePicker", "CirclePicker", "CompactPicker", "GithubPicker",
  "HuePicker", "MaterialPicker", "PhotoshopPicker", "SliderPicker",
  "SwatchesPicker", "TwitterPicker", "AlphaPicker"), label = NULL,
  input_id = NULL, width = NULL, height = NULL)
}
\arguments{
\item{...}{Arguments passed to the picker.}

\item{picker}{Style of the picker.}

\item{label}{Label for the button.}

\item{input_id}{Widget id to retrieve value server-side.}

\item{width, height}{Width and height for the widget.}
}
\description{
Create a color picker among 13 available styles.
}
