% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{compact_picker}
\alias{compact_picker}
\title{Compact Picker Input}
\usage{
compact_picker(input_id, label, color = "#112446", colors = NULL,
  inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{colors}{Vector of colors to include in the widget.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Compact Picker Input
}
\examples{
# See picker in Viewer (or browser)
compact_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {



}
}
