% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{github_picker}
\alias{github_picker}
\title{Github Picker Input}
\usage{
github_picker(input_id, label, color = "#112446", colors = NULL,
  triangle = "top-left", inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{colors}{Vector of colors to include in the widget.}

\item{triangle}{Little arrow above the picker : \code{"hide"}, \code{"top-right"} or \code{"top-left"}.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Github Picker Input
}
\examples{
# See picker in Viewer (or browser)
github_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {



}
}
