% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{photoshop_picker}
\alias{photoshop_picker}
\title{Photoshop Picker Input}
\usage{
photoshop_picker(input_id, label, color = "#112446",
  header = "Color picker", on_accept = NULL, on_cancel = NULL,
  inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{header}{String, Title text.}

\item{on_accept}{Callback for when accept is clicked, use \code{\link[htmlwidgets]{JS}}.}

\item{on_cancel}{Callback for when cancel is clicked, use \code{\link[htmlwidgets]{JS}}.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Photoshop Picker Input
}
\examples{
# See picker in Viewer (or browser)
photoshop_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {



}
}
