% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bg_color.R
\name{set_bg_color}
\alias{set_bg_color}
\title{Change background color of a Shiny app}
\usage{
set_bg_color(color, session = getDefaultReactiveDomain())
}
\arguments{
\item{color}{Hex code color to use as background}

\item{session}{The shiny session within which to call \code{set_bg_color}.}
}
\description{
Change background color of a Shiny app
}
