% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{sketch_picker}
\alias{sketch_picker}
\title{Sketch Picker Input}
\usage{
sketch_picker(input_id, label, color = "#112446", preset_colors = NULL,
  inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{preset_colors}{Vector of colors to be displayed at the bottom of the widget.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Sketch Picker Input
}
\examples{
if (interactive()) {



}
}
