% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{slider_picker}
\alias{slider_picker}
\title{Slider Picker Input}
\usage{
slider_picker(input_id, label, color = "#112446", inline = FALSE,
  width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Slider Picker Input
}
\examples{
# See picker in Viewer (or browser)
slider_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {



}
}
