% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{block_picker}
\alias{block_picker}
\title{Block Picker Input}
\usage{
block_picker(input_id, label, color = "#112446", colors = NULL,
  triangle = "top", inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{colors}{Vector of colors to include in the widget.}

\item{triangle}{Little arrow above the picker : \code{"hide"} or \code{"top"}.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Block Picker Input
}
\examples{
# See picker in Viewer (or browser)
block_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {

  library(shiny)

  ui <- fluidPage(
    tags$h2("Block picker color"),
    block_picker("inline", "Pick!", inline = TRUE),
    verbatimTextOutput("res_inline"),
    block_picker("btn", "Pick!", inline = FALSE),
    verbatimTextOutput("res_btn")
  )

  server <- function(input, output, session) {
    output$res_inline <- renderPrint(input$inline)
    output$res_btn <- renderPrint(input$btn)
  }

  shinyApp(ui, server)

}
}
