% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker.R
\name{color_picker}
\alias{color_picker}
\title{Color Picker Inline}
\usage{
color_picker(color = "#FFFFFF", ..., picker = c("SketchPicker",
  "BlockPicker", "ChromePicker", "CirclePicker", "CompactPicker",
  "GithubPicker", "HuePicker", "MaterialPicker", "PhotoshopPicker",
  "SliderPicker", "SwatchesPicker", "TwitterPicker", "AlphaPicker"),
  input_id = NULL, width = NULL, height = NULL)
}
\arguments{
\item{color}{Default color.}

\item{...}{Additional arguments passed to the picker.}

\item{picker}{Style of the picker.}

\item{input_id}{Widget id to retrieve value server-side.}

\item{width, height}{Width and height for the widget.}
}
\description{
Create a color picker among 13 available styles.
}
\examples{

# May not work weel in RStudio viewer
color_picker(
  inputId = "mypicker",
  picker = "Block",
  colors = c("#EFF3FF", "#C6DBEF", "#9ECAE1",
             "#6BAED6", "#3182BD", "#08519C")
)
}
