% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_picker.R
\name{color_picker}
\alias{color_picker}
\title{Color Picker with Button}
\usage{
color_picker(color = "#112446", ..., picker = c("SketchPicker",
  "BlockPicker", "ChromePicker", "CirclePicker", "CompactPicker",
  "GithubPicker", "HuePicker", "MaterialPicker", "PhotoshopPicker",
  "SliderPicker", "SwatchesPicker", "TwitterPicker", "AlphaPicker"),
  input_id = NULL, label = NULL, inline = TRUE, width = NULL,
  height = NULL)
}
\arguments{
\item{color}{Default color.}

\item{...}{Arguments passed to the picker.}

\item{picker}{Style of the picker.}

\item{input_id}{Widget id to retrieve value server-side.}

\item{label}{Label for the button.}

\item{width, height}{Width and height for the widget.}
}
\description{
Create a color picker among 13 available styles.
}
\examples{

color_picker(
  picker = "Block",
  triangle  = "hide",
  colors = c("#EFF3FF", "#C6DBEF", "#9ECAE1",
             "#6BAED6", "#3182BD", "#08519C", "#112446"),
  label = "Pick a color:"
)
}
