% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{hue_picker}
\alias{hue_picker}
\title{Hue Picker Input}
\usage{
hue_picker(input_id, label, color = "#112446",
  direction = "horizontal", inline = FALSE, width = NULL)
}
\arguments{
\item{input_id}{Id to retrieve the widget value server side.}

\item{label}{Label to be display above the widget or on the button.}

\item{color}{Default color.}

\item{direction}{Direction for the widget: \code{"horizontal"} or \code{"vertical"}.}

\item{inline}{Display the widget inline or when a button is clicked.}

\item{width}{Width for the widget.}
}
\description{
Hue Picker Input
}
\examples{
# See picker in Viewer (or browser)
hue_picker("id", "Pick!", inline = TRUE)

# Shiny example
if (interactive()) {



}
}
